refactor TextInput to make it reusable
